test(flow-client): resolve connection state interference in GWT tests
